replace spec test with unit test in tests/unit_node/http_test.ts
