--- tests/specs/node/request_body_byob_incoming_message/__test__.jsonc
@@ -1,4 +1,0 @@
-{
-  "args": "run -A main.ts",
-  "output": "main.out"
-}
