--- tests/specs/node/request_body_byob_incoming_message/main.out
@@ -1,3 +1,0 @@
-done: false len: 11
-text: hello world
-client got: OK
